Update calculation to make it work
Rewrite function to update the calculation to make it work.
Correctly add the updateDisplay function that shows the calculation on
the display always.
Also add the eventListener to the buttons so they the buttons on the
page work now, therefore are giving the textContent of the buttons as
numbers and operator to the operate function and correctly calculate it